Added comments and renamed files per reviewers' feedback
