+++ ttnn/cpp/ttnn/operations/reduction/generic/device/reduce_op.cpp
@@ -115,7 +115,8 @@ Tensor reduce(
 
     // The single-core HW path uses REDUCE_SCALAR mode, which applies the
     // scaler twice internally (once per dimension).  The host compensates with
-    // sqrt(scaler), but sqrt of a negative number is NaN, so negative scalers
+    // sqrt(scaler) in ReduceSingleCoreHwProgramFactory::create.
+    // However, sqrt of a negative number is NaN, so negative scalers
     // must take the two-step W-then-H path where the scaler is applied once.
     if (is_multicore_hw || (reduce_dim == tt::tt_metal::ReduceOpDim::HW && scaler < 0)) {
         // Multi-core HW reduction: first reduce W, then reduce H on the result
